requirements.txt
langchain
langchain_openai 
langchain_core
python-dotenv
langserve
fastapi
uvicorn
langchain_community
google-generativeai